requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
boto3==1.34.89
botocore==1.34.89
certifi==2024.2.2
cffi==1.16.0
Django==5.0.4
django-minio-storage==0.5.7
django-storages==1.14.2
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
install==1.3.5
jmespath==1.0.1
minio==7.2.5
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1